requirement.txt
PyMuPDF
pytesseract
boto3
openai 
requests
psycopg2
pyjwt
uvicorn
FastAPI
Pillow
pydantic
langchain-community
langgraph
botocore
selenium
beautifulsoup4
python-dotenv
python-multipart